Update ETL to deal with incomplete dates

+++ OpenCitations/ETL/load-open-citations-CSVs.sh
@@ -2,12 +2,11 @@
 set -e
 set -o pipefail
 
-readonly VER=4.2.1
+readonly VER=4.3.0
 
 # Remove the longest `*/` prefix
 readonly SCRIPT_FULL_NAME="${0##*/}"
 max_parallel_jobs=1
-chunk_size=100000
 
 usage() {
   cat <<HEREDOC
@@ -17,12 +16,12 @@ NAME
 
 SYNOPSIS
 
-    $SCRIPT_FULL_NAME [-c] [-j parallel_jobs] [-s chunk_size] [data_directory]
+    $SCRIPT_FULL_NAME [-c] [-j parallel_jobs] [-s batch_max_records] [data_directory]
     $SCRIPT_FULL_NAME -h: display this help
 
 DESCRIPTION
 
-    Load Open Citations CSVs in parallel. Split input files into chunks.
+    Load Open Citations CSVs in parallel. Split input files into batches.
 
     The following options are available:
 
@@ -34,7 +33,7 @@ DESCRIPTION
 
     -j              maximum number of parallel jobs [DEFAULT: $max_parallel_jobs]
 
-    -s              maximum number of lines (citations) per chunk and transaction [DEFAULT: $chunk_size]
+    -s              split and process each CSV in batches up to this maximum number of records
 
 ENVIRONMENT
 
@@ -70,7 +69,7 @@ while getopts cj:s:h OPT; do
     max_parallel_jobs=$OPTARG
     ;;
   s)
-    chunk_size=$OPTARG
+    declare -rx BATCH_SIZE=$OPTARG
     ;;
   *) # -h or `?`: an unknown option
     usage
@@ -97,7 +96,7 @@ readonly SCRIPT_DIR=${SCRIPT_FILENAME%/*}
 readonly ABSOLUTE_SCRIPT_DIR=$(cd "${SCRIPT_DIR}" && pwd)
 
 ########################################################################################################################
-# Loads a single header-less CSV chunk of Open Citations.
+# Loads a single header-less CSV batch of Open Citations.
 # Executed in a subshell by `parallel`.
 #
 # Arguments:
@@ -106,17 +105,23 @@ readonly ABSOLUTE_SCRIPT_DIR=$(cd "${SCRIPT_DIR}" && pwd)
 load_csv() {
   set -e
   set -o pipefail
-  local absolute_file_path=$1
-  echo "Processing ${absolute_file_path} ..."
+  local -r ABSOLUTE_FILE_PATH=$1
+  echo "Processing ${ABSOLUTE_FILE_PATH} ..."
+
+  if [[ $BATCH_SIZE ]]; then
+    local -r HEADERS=OFF
+  else
+    local -r HEADERS="ON MATCH"
+  fi
   # language=PostgresPLSQL
   psql -v ON_ERROR_STOP=on <<HEREDOC
       COPY stg_open_citations (oci, citing, cited, creation, timespan, journal_sc, author_sc)
-      FROM '${absolute_file_path}' (FORMAT CSV, HEADER OFF);
+      FROM '${ABSOLUTE_FILE_PATH}' (FORMAT CSV, HEADER $HEADERS);
 HEREDOC
 
-  echo "Loaded ${absolute_file_path}"
+  echo "Loaded ${ABSOLUTE_FILE_PATH}"
   if [[ $REMOVE_LOADED ]]; then
-    rm -v "${absolute_file_path}"
+    rm -v "${ABSOLUTE_FILE_PATH}"
   fi
 }
 export -f load_csv
@@ -126,26 +131,29 @@ echo "Starting data load: appending all records to existing Open Citations data.
 psql -f "$ABSOLUTE_SCRIPT_DIR/pre_processing.sql"
 
 cd "$DATA_DIR"
-if [[ ! -d chunks ]]; then
-  mkdir -p chunks
-  # shellcheck disable=SC2016 # `--tagstring` tokens are expanded by GNU `parallel`
-  find . -maxdepth 1 -name '*.csv' -type f -print0 | parallel -0 -j "$max_parallel_jobs" --halt soon,fail=1 \
-    --verbose --line-buffer --tagstring '|job #{#} of {= $_=total_jobs() =} slot #{%}|' \
-      "tail -n +2 {} | split --lines=$chunk_size --numeric-suffixes=1 --elide-empty-files --additional-suffix=.csv \
-        - chunks/{}.part"
+if [[ $BATCH_SIZE ]]; then
+  if [[ ! -d batches ]]; then
+    mkdir -p batches
+    # shellcheck disable=SC2016 # `--tagstring` tokens are expanded by GNU `parallel`
+    find . -maxdepth 1 -name '*.csv' -type f -print0 | parallel -0 -j "$max_parallel_jobs" --halt soon,fail=1 \
+        --verbose --line-buffer --tagstring '|job #{#} of {= $_=total_jobs() =} slot #{%}|' \
+        "tail -n +2 {} | split --lines=$BATCH_SIZE --numeric-suffixes=1 --elide-empty-files --additional-suffix=.csv - \
+            batches/{}.part"
+  fi
+  cd batches
 fi
-cd chunks
 
-# Load chunks. If load fails, the process should restart here on the remaining chunks.
+# Load files. If the load fails, the process can nbe resumed here on the remaining files.
 # Piping to `parallel` is done by design here to handle a large number of files potentially
 # shellcheck disable=SC2016 # `--tagstring` tokens are expanded by GNU `parallel`
 find ~+ -maxdepth 1 -type f -name '*.csv' -print0 | parallel -0 -j "$max_parallel_jobs" --halt soon,fail=1 \
-  --line-buffer --tagstring '|job #{#} of {= $_=total_jobs() =} slot #{%}|' load_csv '{}'
+    --line-buffer --tagstring '|job #{#} of {= $_=total_jobs() =} slot #{%}|' load_csv '{}'
 cd "$DATA_DIR"
 
 # Successfully loaded all input files
-rmdir chunks
-if [[ $REMOVE_LOADED ]]; then
+rm -rf batches
+if [[ $REMOVE_LOADED && $BATCH_SIZE ]]; then
+  # Remove original files (they were split into batches and left over)
   rm -fv -- *.csv
 fi
